UI prevents placing towers that prevent pathfinding [#6 state:resolved]
However, every time the mouse moves so much as a pixel, it does full pathfinding,
which really reduces the responsiveness and gives it a sluggish
feel, which can really kill a game.

We could improve this by only checking for pathfinding legality when the
user attempts to place the tower, or we could only check a block once.
This could be done specifically for pathfinding, or the is_legal method
of a state could be called only when the currently selected block changes
rather than at the pixel level.