Sound effects!  Try launching a Small Missile!
This was a bigger feature than I expected, and rather tricky to
do it right and with the performance characteristics that
we want for a tight, responsive game.  Still has a ways to
go, and this is just a proof of concept.  Not even sure
that sound effects will make it more fun.

I ended up using a hybrid model of soundmanager2 and some
custom javascript embedding as a fallback.  If the flash
croaks (by default Flash and Javascript aren't allowed to talk
to one another in some situations on some platforms) then
it falls back to an <embed> element.  Since we're using
an mp3 to ensure a quick load time, the <embed> only works
in browsers that support mp3 (read: Safari, perhaps IE?).

Too much to talk about in a commit message, expect a wiki
article.