Added some logging statements to track down Small Missile bug.
The problem is that no creep is ever determined to be within the
radius of the missile.

It seems as though this is due to the <= method being broken... which
is deeply strange.  Relevant logs:

radius: 125
radius type: number
distance: 34.44927389621228
distance type: number
distance <= radius: false

Am I losing my mind here?

Also, I think it'd be best to change Small Missile to Bomb to not
overlap verbally with Missile Towers.