Fixed underline problem in deploy, cleaned html
[#20 state:resolved]
BeautifulSoup isn't beautifying our html any more, but I've
gone through and validated our html as it stands (besides the
<canvas> element), and I'll take up that responsibility.

I also wrote a simple deploy script.  If it gets much more
complicated, we should look at rake.  Sure it adds another
dependency, but it's *so* nice for doing this sort of thing.