Changed some colors around for visual distinctiveness
The bomb radius is now a menacing black.  Or a bureaucratic gray,
I guess.

Laser towers are now a sadder green, but you can see them in water.
I don't do this because it is a good color for them, rather, I do it
to make sure that we come to a better solution before 1.0 launch.